+++ ethergo/submitter/submitter.go
@@ -11,6 +11,8 @@ import (
 	"sync"
 	"time"
 
+	"github.com/puzpuzpuz/xsync/v2"
+
 	"github.com/ethereum/go-ethereum/accounts/abi/bind"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/core/types"
@@ -64,6 +66,8 @@ type txSubmitterImpl struct {
 	// to prevent memory leaks, this has a buffer of 1.
 	// callers adding to this channel should not block.
 	retryNow chan bool
+	// lastGasBlockCache is used to cache the last gas block for a given chain. A new block should still be fetched, if possible.
+	lastGasBlockCache *xsync.MapOf[int, *types.Header]
 	// config is the config for the transaction submitter.
 	config config.IConfig
 }
@@ -78,14 +82,15 @@ type ClientFetcher interface {
 // NewTransactionSubmitter creates a new transaction submitter.
 func NewTransactionSubmitter(metrics metrics.Handler, signer signer.Signer, fetcher ClientFetcher, db db.Service, config config.IConfig) TransactionSubmitter {
 	return &txSubmitterImpl{
-		db:        db,
-		config:    config,
-		metrics:   metrics,
-		signer:    signer,
-		fetcher:   fetcher,
-		nonceMux:  mapmutex.NewStringerMapMutex(),
-		statusMux: mapmutex.NewStringMapMutex(),
-		retryNow:  make(chan bool, 1),
+		db:                db,
+		config:            config,
+		metrics:           metrics,
+		signer:            signer,
+		fetcher:           fetcher,
+		nonceMux:          mapmutex.NewStringerMapMutex(),
+		statusMux:         mapmutex.NewStringMapMutex(),
+		retryNow:          make(chan bool, 1),
+		lastGasBlockCache: xsync.NewIntegerMapOf[int, *types.Header](),
 	}
 }
 
@@ -379,8 +384,7 @@ func (t *txSubmitterImpl) setGasPrice(ctx context.Context, client client.EVM,
 
 	//nolint: nestif
 	if prevTx != nil {
-		// TODO: cache
-		gasBlock, err := t.getGasBlock(ctx, client)
+		gasBlock, err := t.getGasBlock(ctx, client, chainID)
 		if err != nil {
 			span.AddEvent("could not get gas block", trace.WithAttributes(attribute.String("error", err.Error())))
 			return err
@@ -419,7 +423,7 @@ func (t *txSubmitterImpl) setGasPrice(ctx context.Context, client client.EVM,
 }
 
 // getGasBlock gets the gas block for the given chain.
-func (t *txSubmitterImpl) getGasBlock(ctx context.Context, chainClient client.EVM) (gasBlock *types.Header, err error) {
+func (t *txSubmitterImpl) getGasBlock(ctx context.Context, chainClient client.EVM, chainID int) (gasBlock *types.Header, err error) {
 	ctx, span := t.metrics.Tracer().Start(ctx, "submitter.getGasBlock")
 	defer func() {
 		metrics.EndSpanWithErr(span, err)
@@ -434,9 +438,22 @@ func (t *txSubmitterImpl) getGasBlock(ctx context.Context, chainClient client.EV
 		return nil
 	}, retry.WithMin(time.Millisecond*50), retry.WithMax(time.Second*3), retry.WithMaxAttempts(4))
 
+	// if we can't get the current gas block, attempt to load it from the cache
 	if err != nil {
-		return nil, fmt.Errorf("could not get gas block: %w", err)
+		var ok bool
+		gasBlock, ok = t.lastGasBlockCache.Load(chainID)
+		if ok {
+			span.AddEvent("could not get gas block; using cached value", trace.WithAttributes(
+				attribute.String("error", err.Error()),
+				attribute.String("blockNumber", bigPtrToString(gasBlock.Number)),
+			))
+		} else {
+			return nil, fmt.Errorf("could not get gas block: %w", err)
+		}
 	}
+
+	// cache the latest gas block
+	t.lastGasBlockCache.Store(chainID, gasBlock)
 
 	return gasBlock, nil
 }