+++ packages/synapse-interface/utils/hasOnlyZeroes.ts
@@ -1,0 +1,3 @@
+export const hasOnlyZeroes = (input: string): boolean => {
+  return /^0+(\.0+)?$/.test(input)
+}